+++ tests/e2e_test.go
@@ -71,4 +71,7 @@ var _ = ginkgo.Describe("[AWM Relayer Integration Tests", func() {
 	ginkgo.It("Teleporter Registry", func() {
 		TeleporterRegistry(localNetworkInstance)
 	})
+	ginkgo.It("Shared Database", func() {
+		SharedDatabaseAccess(localNetworkInstance)
+	})
 })
