+++ tests/teleporter_registry.go
@@ -81,6 +81,7 @@ func TeleporterRegistry(network interfaces.LocalNetwork) {
 	//
 	relayerConfig := testUtils.CreateDefaultRelayerConfig(
 		[]interfaces.SubnetTestInfo{cChainInfo},
+		[]interfaces.SubnetTestInfo{cChainInfo},
 		teleporterContractAddress,
 		fundedAddress,
 		relayerKey,
@@ -94,7 +95,7 @@ func TeleporterRegistry(network interfaces.LocalNetwork) {
 			DestinationAddress:      cChainInfo.TeleporterRegistryAddress.Hex(),
 		},
 	}
-	relayerConfigPath := writeRelayerConfig(relayerConfig)
+	relayerConfigPath := testUtils.WriteRelayerConfig(relayerConfig, testUtils.DefaultRelayerCfgFname)
 	//
 	// Run the Relayer. On startup, we should deliver the message provided in the config
 	//
