+++ tests/utils/utils.go
@@ -8,18 +8,29 @@ import (
 	"context"
 	"crypto/ecdsa"
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
 	"math/big"
 	"os"
 	"os/exec"
 	"strings"
 
+	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/logging"
+	avalancheWarp "github.com/ava-labs/avalanchego/vms/platformvm/warp"
+	warpPayload "github.com/ava-labs/avalanchego/vms/platformvm/warp/payload"
 	"github.com/ava-labs/awm-relayer/config"
 	offchainregistry "github.com/ava-labs/awm-relayer/messages/off-chain-registry"
+	"github.com/ava-labs/subnet-evm/accounts/abi/bind"
+	"github.com/ava-labs/subnet-evm/core/types"
+	"github.com/ava-labs/subnet-evm/precompile/contracts/warp"
+	predicateutils "github.com/ava-labs/subnet-evm/predicate"
+	subnetevmutils "github.com/ava-labs/subnet-evm/utils"
+	teleportermessenger "github.com/ava-labs/teleporter/abi-bindings/go/Teleporter/TeleporterMessenger"
 	"github.com/ava-labs/teleporter/tests/interfaces"
 	"github.com/ava-labs/teleporter/tests/utils"
 	teleporterTestUtils "github.com/ava-labs/teleporter/tests/utils"
+	teleporterUtils "github.com/ava-labs/teleporter/utils/teleporter-utils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/log"
@@ -27,7 +38,9 @@ import (
 )
 
 // Write the test database to /tmp since the data is not needed after the test
-var storageLocation = fmt.Sprintf("%s/.awm-relayer-storage", os.TempDir())
+var StorageLocation = fmt.Sprintf("%s/.awm-relayer-storage", os.TempDir())
+
+const DefaultRelayerCfgFname = "relayer-config.json"
 
 func BuildAndRunRelayerExecutable(ctx context.Context, relayerConfigPath string) context.CancelFunc {
 	// Build the awm-relayer binary
@@ -83,7 +96,8 @@ func ReadHexTextFile(filename string) string {
 
 // Constructs a relayer config with all subnets as sources and destinations
 func CreateDefaultRelayerConfig(
-	subnetsInfo []interfaces.SubnetTestInfo,
+	sourceSubnetsInfo []interfaces.SubnetTestInfo,
+	destinationSubnetsInfo []interfaces.SubnetTestInfo,
 	teleporterContractAddress common.Address,
 	fundedAddress common.Address,
 	relayerKey *ecdsa.PrivateKey,
@@ -92,9 +106,9 @@ func CreateDefaultRelayerConfig(
 		"Setting up relayer config",
 	)
 	// Construct the config values for each subnet
-	sources := make([]*config.SourceBlockchain, len(subnetsInfo))
-	destinations := make([]*config.DestinationBlockchain, len(subnetsInfo))
-	for i, subnetInfo := range subnetsInfo {
+	sources := make([]*config.SourceBlockchain, len(sourceSubnetsInfo))
+	destinations := make([]*config.DestinationBlockchain, len(destinationSubnetsInfo))
+	for i, subnetInfo := range sourceSubnetsInfo {
 		host, port, err := teleporterTestUtils.GetURIHostAndPort(subnetInfo.NodeURIs[0])
 		Expect(err).Should(BeNil())
 
@@ -121,6 +135,19 @@ func CreateDefaultRelayerConfig(
 			},
 		}
 
+		log.Info(
+			"Creating relayer config for source subnet",
+			"subnetID", subnetInfo.SubnetID.String(),
+			"blockchainID", subnetInfo.BlockchainID.String(),
+			"host", host,
+			"port", port,
+		)
+	}
+
+	for i, subnetInfo := range destinationSubnetsInfo {
+		host, port, err := teleporterTestUtils.GetURIHostAndPort(subnetInfo.NodeURIs[0])
+		Expect(err).Should(BeNil())
+
 		destinations[i] = &config.DestinationBlockchain{
 			SubnetID:          subnetInfo.SubnetID.String(),
 			BlockchainID:      subnetInfo.BlockchainID.String(),
@@ -130,7 +157,7 @@ func CreateDefaultRelayerConfig(
 		}
 
 		log.Info(
-			"Creating relayer config for subnet",
+			"Creating relayer config for destination subnet",
 			"subnetID", subnetInfo.SubnetID.String(),
 			"blockchainID", subnetInfo.BlockchainID.String(),
 			"host", host,
@@ -140,21 +167,17 @@ func CreateDefaultRelayerConfig(
 
 	return config.Config{
 		LogLevel:               logging.Info.LowerString(),
-		PChainAPIURL:           subnetsInfo[0].NodeURIs[0],
-		InfoAPIURL:             subnetsInfo[0].NodeURIs[0],
-		StorageLocation:        RelayerStorageLocation(),
+		PChainAPIURL:           sourceSubnetsInfo[0].NodeURIs[0],
+		InfoAPIURL:             sourceSubnetsInfo[0].NodeURIs[0],
+		StorageLocation:        StorageLocation,
 		ProcessMissedBlocks:    false,
 		SourceBlockchains:      sources,
 		DestinationBlockchains: destinations,
 	}
 }
 
-func RelayerStorageLocation() string {
-	return storageLocation
-}
-
 func ClearRelayerStorage() error {
-	return os.RemoveAll(storageLocation)
+	return os.RemoveAll(StorageLocation)
 }
 
 func FundRelayers(
@@ -172,4 +195,205 @@ func FundRelayers(
 		)
 		utils.SendTransactionAndWaitForSuccess(ctx, subnetInfo, fundRelayerTx)
 	}
+}
+
+func SendBasicTeleporterMessage(
+	ctx context.Context,
+	source interfaces.SubnetTestInfo,
+	destination interfaces.SubnetTestInfo,
+	fundedKey *ecdsa.PrivateKey,
+	fundedAddress common.Address,
+) (*types.Receipt, teleportermessenger.TeleporterMessage, ids.ID) {
+	input := teleportermessenger.TeleporterMessageInput{
+		DestinationBlockchainID: destination.BlockchainID,
+		DestinationAddress:      fundedAddress,
+		FeeInfo: teleportermessenger.TeleporterFeeInfo{
+			FeeTokenAddress: fundedAddress,
+			Amount:          big.NewInt(0),
+		},
+		RequiredGasLimit:        big.NewInt(1),
+		AllowedRelayerAddresses: []common.Address{},
+		Message:                 []byte{1, 2, 3, 4},
+	}
+
+	// Send a transaction to the Teleporter contract
+	log.Info(
+		"Sending teleporter transaction",
+		"sourceBlockchainID", source.BlockchainID,
+		"destinationBlockchainID", destination.BlockchainID,
+	)
+	receipt, teleporterMessageID := teleporterTestUtils.SendCrossChainMessageAndWaitForAcceptance(
+		ctx,
+		source,
+		destination,
+		input,
+		fundedKey,
+	)
+	sendEvent, err := teleporterTestUtils.GetEventFromLogs(receipt.Logs, source.TeleporterMessenger.ParseSendCrossChainMessage)
+	Expect(err).Should(BeNil())
+
+	return receipt, sendEvent.Message, teleporterMessageID
+}
+
+func RelayBasicMessage(
+	ctx context.Context,
+	source interfaces.SubnetTestInfo,
+	destination interfaces.SubnetTestInfo,
+	teleporterContractAddress common.Address,
+	fundedKey *ecdsa.PrivateKey,
+	fundedAddress common.Address,
+) {
+	newHeadsDest := make(chan *types.Header, 10)
+	sub, err := destination.WSClient.SubscribeNewHead(ctx, newHeadsDest)
+	Expect(err).Should(BeNil())
+	defer sub.Unsubscribe()
+
+	_, teleporterMessage, teleporterMessageID := SendBasicTeleporterMessage(
+		ctx,
+		source,
+		destination,
+		fundedKey,
+		fundedAddress,
+	)
+
+	log.Info("Waiting for new block confirmation")
+	newHead := <-newHeadsDest
+	blockNumber := newHead.Number
+	log.Info(
+		"Received new head",
+		"height", blockNumber.Uint64(),
+		"hash", newHead.Hash(),
+	)
+	block, err := destination.RPCClient.BlockByNumber(ctx, blockNumber)
+	Expect(err).Should(BeNil())
+	log.Info(
+		"Got block",
+		"blockHash", block.Hash(),
+		"blockNumber", block.NumberU64(),
+		"transactions", block.Transactions(),
+		"numTransactions", len(block.Transactions()),
+		"block", block,
+	)
+	accessLists := block.Transactions()[0].AccessList()
+	Expect(len(accessLists)).Should(Equal(1))
+	Expect(accessLists[0].Address).Should(Equal(warp.Module.Address))
+
+	// Check the transaction storage key has warp message we're expecting
+	storageKeyHashes := accessLists[0].StorageKeys
+	packedPredicate := subnetevmutils.HashSliceToBytes(storageKeyHashes)
+	predicateBytes, err := predicateutils.UnpackPredicate(packedPredicate)
+	Expect(err).Should(BeNil())
+	receivedWarpMessage, err := avalancheWarp.ParseMessage(predicateBytes)
+	Expect(err).Should(BeNil())
+
+	// Check that the transaction has successful receipt status
+	txHash := block.Transactions()[0].Hash()
+	receipt, err := destination.RPCClient.TransactionReceipt(ctx, txHash)
+	Expect(err).Should(BeNil())
+	Expect(receipt.Status).Should(Equal(types.ReceiptStatusSuccessful))
+
+	// Check that the transaction emits ReceiveCrossChainMessage
+	receiveEvent, err := teleporterTestUtils.GetEventFromLogs(receipt.Logs, destination.TeleporterMessenger.ParseReceiveCrossChainMessage)
+	Expect(err).Should(BeNil())
+	Expect(receiveEvent.SourceBlockchainID[:]).Should(Equal(source.BlockchainID[:]))
+	Expect(receiveEvent.MessageID[:]).Should(Equal(teleporterMessageID[:]))
+
+	//
+	// Validate Received Warp Message Values
+	//
+	log.Info("Validating received warp message")
+	Expect(receivedWarpMessage.SourceChainID).Should(Equal(source.BlockchainID))
+	addressedPayload, err := warpPayload.ParseAddressedCall(receivedWarpMessage.Payload)
+	Expect(err).Should(BeNil())
+
+	// Check that the teleporter message is correct
+	// We don't validate the entire message, since the message receipts
+	// are populated by the Teleporter contract
+	receivedTeleporterMessage, err := teleportermessenger.UnpackTeleporterMessage(addressedPayload.Payload)
+	Expect(err).Should(BeNil())
+
+	receivedMessageID, err := teleporterUtils.CalculateMessageID(teleporterContractAddress, source.BlockchainID, destination.BlockchainID, teleporterMessage.MessageNonce)
+	Expect(err).Should(BeNil())
+	Expect(receivedMessageID).Should(Equal(teleporterMessageID))
+	Expect(receivedTeleporterMessage.OriginSenderAddress).Should(Equal(teleporterMessage.OriginSenderAddress))
+	receivedDestinationID, err := ids.ToID(receivedTeleporterMessage.DestinationBlockchainID[:])
+	Expect(err).Should(BeNil())
+	Expect(receivedDestinationID).Should(Equal(destination.BlockchainID))
+	Expect(receivedTeleporterMessage.DestinationAddress).Should(Equal(teleporterMessage.DestinationAddress))
+	Expect(receivedTeleporterMessage.RequiredGasLimit.Uint64()).Should(Equal(teleporterMessage.RequiredGasLimit.Uint64()))
+	Expect(receivedTeleporterMessage.Message).Should(Equal(teleporterMessage.Message))
+}
+
+func WriteRelayerConfig(relayerConfig config.Config, fname string) string {
+	data, err := json.MarshalIndent(relayerConfig, "", "\t")
+	Expect(err).Should(BeNil())
+
+	f, err := os.CreateTemp(os.TempDir(), fname)
+	Expect(err).Should(BeNil())
+
+	_, err = f.Write(data)
+	Expect(err).Should(BeNil())
+	relayerConfigPath := f.Name()
+
+	log.Info("Created awm-relayer config", "configPath", relayerConfigPath, "config", string(data))
+	return relayerConfigPath
+}
+
+func TriggerProcessMissedBlocks(
+	ctx context.Context,
+	sourceSubnetInfo interfaces.SubnetTestInfo,
+	destinationSubnetInfo interfaces.SubnetTestInfo,
+	currRelayerCleanup context.CancelFunc,
+	currrentRelayerConfig config.Config,
+	fundedAddress common.Address,
+	fundedKey *ecdsa.PrivateKey,
+) {
+	// First, make sure the relayer is stopped
+	currRelayerCleanup()
+
+	// Subscribe to the destination chain
+	newHeads := make(chan *types.Header, 10)
+	sub, err := destinationSubnetInfo.WSClient.SubscribeNewHead(ctx, newHeads)
+	Expect(err).Should(BeNil())
+	defer sub.Unsubscribe()
+
+	// Send three Teleporter messages from subnet A to subnet B
+	log.Info("Sending three Teleporter messages from subnet A to subnet B")
+	_, _, id1 := SendBasicTeleporterMessage(ctx, sourceSubnetInfo, destinationSubnetInfo, fundedKey, fundedAddress)
+	_, _, id2 := SendBasicTeleporterMessage(ctx, sourceSubnetInfo, destinationSubnetInfo, fundedKey, fundedAddress)
+	_, _, id3 := SendBasicTeleporterMessage(ctx, sourceSubnetInfo, destinationSubnetInfo, fundedKey, fundedAddress)
+
+	currHeight, err := sourceSubnetInfo.RPCClient.BlockNumber(ctx)
+	Expect(err).Should(BeNil())
+	log.Info("Current block height", "height", currHeight)
+
+	// Configure the relayer such that it will only process the last of the three messages sent above.
+	// The relayer DB stores the height of the block *before* the first message, so by setting the
+	// ProcessHistoricalBlocksFromHeight to the block height of the *third* message, we expect the relayer to skip
+	// the first two messages on startup, but process the third.
+	modifiedRelayerConfig := currrentRelayerConfig
+	modifiedRelayerConfig.SourceBlockchains[0].ProcessHistoricalBlocksFromHeight = currHeight
+	modifiedRelayerConfig.ProcessMissedBlocks = true
+	relayerConfigPath := WriteRelayerConfig(modifiedRelayerConfig, DefaultRelayerCfgFname)
+
+	log.Info("Starting the relayer")
+	relayerCleanup := BuildAndRunRelayerExecutable(ctx, relayerConfigPath)
+	defer relayerCleanup()
+	log.Info("Waiting for a new block confirmation on subnet B")
+	<-newHeads
+	delivered1, err := destinationSubnetInfo.TeleporterMessenger.MessageReceived(
+		&bind.CallOpts{}, id1,
+	)
+	Expect(err).Should(BeNil())
+	delivered2, err := destinationSubnetInfo.TeleporterMessenger.MessageReceived(
+		&bind.CallOpts{}, id2,
+	)
+	Expect(err).Should(BeNil())
+	delivered3, err := destinationSubnetInfo.TeleporterMessenger.MessageReceived(
+		&bind.CallOpts{}, id3,
+	)
+	Expect(err).Should(BeNil())
+	Expect(delivered1).Should(BeFalse())
+	Expect(delivered2).Should(BeFalse())
+	Expect(delivered3).Should(BeTrue())
 }
